requirements.txt
selenium==4.2.0
discord
discord.py
python-dotenv
gTTS
langdetect